Fix that Context were used in multiple threads at the same time
For each request, we would assign context.CurrentUser with the current
user, but when using this multithreaded, all threads would use the same
context and hence fight for who will set the correct User.

Split context and Username. Username will be provided to the handles via
Session (that is created for that request). They still share same
context (and thereby repositories, so we got a problem there)